Move crypto test out of Secrets route
