+++ server/routes/v1/index.js
@@ -2,6 +2,7 @@ var router = require('express').Router()
 
 router.use('/secret', require('./secret'))
 router.use('/passphrase', require('./passphrase'))
+router.use('/test', require('./crypto_test'))
 
 router.get('/', (req, res) => {
   res.send({'status': 200})
